Move label component to forms folder
