+++ site/gatsby-site/src/components/forms/IncidentReportForm.js
@@ -4,7 +4,7 @@ import { Form, Button, Spinner } from 'react-bootstrap';
 import { useFormik } from 'formik';
 import * as Yup from 'yup';
 import Link from 'components/ui/Link';
-import TextInputGroup from 'components/TextInputGroup';
+import TextInputGroup from 'components/forms/TextInputGroup';
 import useToastContext, { SEVERITY } from '../../hooks/useToast';
 import { format, parseISO } from 'date-fns';
 import { dateRegExp } from 'utils/date';
@@ -14,7 +14,7 @@ import 'react-bootstrap-typeahead/css/Typeahead.css';
 import styled from 'styled-components';
 import { graphql, useStaticQuery } from 'gatsby';
 import * as POP_OVERS from '../PopOvers';
-import Label from '../Label';
+import Label from './Label';
 
 // set in form //
 // * title: "title of the report" # (string) The title of the report that is indexed.
